implement first draft of the telemetry system
We want to have a flexible and relatively efficient logging and
telemetry system so that we can have insign into what the application is
doing.
This commit adds a first pass version. The telemetry system in this
commit is flexible and powerful, however it is not yet complete (doesn't
log file/line/function, logs to stdout/stdlog/stderr only, not yet
pluggable, and if a telemetry item is enabled then calling it is not
particularly efficient (lots of string processing)).
We will want to improve on this, but its a starting point to get the
ball rolling and will help determine what the API and implementation
requiremnts for the final version will be.